Switch to use pdata to jaeger translator, remove copied code
