+++ go.mod
@@ -4,11 +4,13 @@ go 1.13
 
 require (
 	github.com/gogo/protobuf v1.3.2
-	github.com/golang/protobuf v1.4.3
-	github.com/google/go-cmp v0.5.4 // indirect
-	github.com/jaegertracing/jaeger v1.22.0
-	github.com/opentracing/opentracing-go v1.2.0 // indirect
+	github.com/golang/protobuf v1.5.2
+	github.com/jaegertracing/jaeger v1.30.0
+	github.com/open-telemetry/opentelemetry-collector-contrib/pkg/translator/jaeger v0.44.0
 	github.com/stretchr/testify v1.7.0
 	go.opencensus.io v0.23.0
-	google.golang.org/grpc v1.33.2
+	go.opentelemetry.io/collector/model v0.44.0
+	go.opentelemetry.io/proto/otlp v0.7.0
+	google.golang.org/grpc v1.44.0
+	google.golang.org/protobuf v1.27.1
 )
